Move `detect` command to `results issues`
The command `detect` is used for detecting KCIDB issues for
builds/boots. Move it to the command group `results issues`.
