+++ kcidev/subcommands/results/__init__.py
@@ -4,10 +4,14 @@ import sys
 from functools import wraps
 
 import click
+from tabulate import tabulate
 
+from kcidev.libs.common import kci_msg_red
 from kcidev.libs.dashboard import (
+    dashboard_fetch_boot_issues,
     dashboard_fetch_boots,
     dashboard_fetch_build,
+    dashboard_fetch_build_issues,
     dashboard_fetch_builds,
     dashboard_fetch_commits_history,
     dashboard_fetch_summary,
@@ -379,6 +383,207 @@ def compare(origin, giturl, branch, latest, commits, use_json):
         click.echo("Error: Provide either --latest flag or exactly 2 commit hashes")
         raise click.Abort()
 
+
+def get_issues(ctx, origin, item_type, giturl, branch, commit, tree_name, arch):
+    """Get KCIDB issues for builds/boots"""
+    try:
+        if item_type == "builds":
+            results_cmd = builds
+            dashboard_func = dashboard_fetch_build_issues
+        elif item_type == "boots":
+            results_cmd = boots
+            dashboard_func = dashboard_fetch_boot_issues
+        else:
+            kci_msg_red("Please specify 'builds' or 'boots' as items type")
+            return []
+
+        dashboard_items = ctx.invoke(
+            results_cmd,
+            origin=origin,
+            giturl=giturl,
+            branch=branch,
+            commit=commit,
+            status="all",
+            count=True,
+            verbose=False,
+            arch=arch,
+        )
+        final_stats = []
+        for item in dashboard_items:
+            # Exclude passed builds/boots
+            if item["status"] == "PASS":
+                continue
+            item_id = item["id"]
+            try:
+                issues = dashboard_func(item_id, False)
+                issue_ids = []
+                for issue in issues:
+                    issue_ids.append(issue["id"])
+                final_stats.append([f"{tree_name}/{branch}", item_id, issue_ids])
+            except click.ClickException as e:
+                if e.message in (
+                    "No issues were found for this build",
+                    "No issues were found for this test",
+                ):
+                    final_stats.append([f"{tree_name}/{branch}", item_id, []])
+        return final_stats
+    except click.Abort:
+        kci_msg_red(
+            f"{tree_name}/{branch}: Aborted while fetching dashboard builds/boots"
+        )
+        return []
+    except click.ClickException as e:
+        kci_msg_red(f"{tree_name}/{branch}: {e.message}")
+        return []
+
+
+def get_issues_for_specific_item(item_type, item_id):
+    """Get KCIDB issues for a specific build/boot matching provided ID"""
+    try:
+        if item_type == "builds":
+            dashboard_func = dashboard_fetch_build_issues
+        elif item_type == "boots":
+            dashboard_func = dashboard_fetch_boot_issues
+        else:
+            kci_msg_red("Please specify 'builds' or 'boots' as items type")
+            return []
+        issues = dashboard_func(item_id, False)
+        issue_ids = []
+        for issue in issues:
+            issue_ids.append(issue["id"])
+        stats = [[item_id, issue_ids]]
+        return stats
+    except click.ClickException as e:
+        if e.message == "No issues were found for this build":
+            stats = [[item_id, []]]
+            return stats
+        return None
+
+
+def print_stats(data, headers, max_col_width, table_fmt):
+    """Print build statistics in tabular format"""
+    print("Creating a stats report...")
+    print(
+        tabulate(data, headers=headers, maxcolwidths=max_col_width, tablefmt=table_fmt)
+    )
+
+
+@click.command(
+    name="detect",
+    help="""Detect KCIDB issues for builds and boots
+
+\b
+The command is used to fetch KCIDB issues associated with builds and boots.
+Provide `--builds` and `--boots` option for builds issue detection and boots
+issue detection respectively.
+`--all-checkouts` option can be used to fetch KCIDB issues for all the
+failed and inconclusive builds/boots from all available trees on the dashboard.
+Issues for a single build/boot can be retrieved by providing `--id`
+option.
+
+\b
+Examples:
+  # Detect build issues
+  kci-dev issues detect --builds --id <build-id>
+  kci-dev issues detect --builds --all-checkouts --days <number-of-days> --origin <origin>
+  # Detect boot issues
+  kci-dev issues detect --boots --id <boot-id>
+  kci-dev issues detect --boots --all-checkouts --days <number-of-days> --origin <origin>
+""",
+)
+@click.option(
+    "--origin",
+    help="Select KCIDB origin",
+    default="maestro",
+)
+@click.option(
+    "--builds",
+    is_flag=True,
+    help="Fetch KCIDB issues for builds",
+)
+@click.option(
+    "--boots",
+    is_flag=True,
+    help="Fetch KCIDB issues for boots",
+)
+@click.option(
+    "--id",
+    "item_id",
+    help="Build/boot id to get issues for",
+)
+@click.option(
+    "--all-checkouts",
+    is_flag=True,
+    help="Fetch KCIDB issues for all failed/inconclusive builds/boots of all available checkouts",
+)
+@click.option("--arch", help="Filter by arch")
+@click.option(
+    "--days",
+    help="Provide a period of time in days to get results for",
+    type=int,
+    default="7",
+)
+@click.pass_context
+def detect(
+    ctx,
+    origin,
+    builds,
+    boots,
+    item_id,
+    all_checkouts,
+    arch,
+    days,
+):
+
+    if not (builds or boots):
+        raise click.UsageError("Provide --builds or --boots to fetch issues for")
+
+    if builds and boots:
+        raise click.UsageError("Specify only one option from --builds and --boots")
+
+    item_type = "builds" if builds else "boots"
+
+    if not (all_checkouts or item_id):
+        raise click.UsageError("Provide --all-checkouts or --id")
+
+    print("Fetching KCIDB issues...")
+    if all_checkouts:
+        if item_id:
+            raise click.UsageError("Cannot use --all-checkouts with --id")
+        final_stats = []
+        trees_list = ctx.invoke(trees, origin=origin, days=days, verbose=False)
+        for tree in trees_list:
+            giturl = tree["git_repository_url"]
+            branch = tree["git_repository_branch"]
+            commit = tree["git_commit_hash"]
+            tree_name = tree["tree_name"]
+            stats = get_issues(
+                ctx, origin, item_type, giturl, branch, commit, tree_name, arch
+            )
+            final_stats.extend(stats)
+        if final_stats:
+            headers = [
+                "tree/branch",
+                "ID",
+                "Issues",
+            ]
+            max_col_width = [None, None, None]
+            table_fmt = "simple_grid"
+            print_stats(final_stats, headers, max_col_width, table_fmt)
+
+    elif item_id:
+        stats = get_issues_for_specific_item(item_type, item_id)
+        if stats:
+            headers = [
+                "ID",
+                "Issues",
+            ]
+            max_col_width = [None, None]
+            table_fmt = "simple_grid"
+            print_stats(stats, headers, max_col_width, table_fmt)
+
+
+issues.add_command(detect)
 
 if __name__ == "__main__":
     main_kcidev()
